all: Redefine ttnpb.Identifiers
