+++ pkg/ttnpb/identifiers.go
@@ -16,7 +16,6 @@ package ttnpb
 
 import (
 	"context"
-	"fmt"
 
 	"go.thethings.network/lorawan-stack/pkg/errors"
 	"go.thethings.network/lorawan-stack/pkg/types"
@@ -56,153 +55,11 @@ func (ids UserIdentifiers) IsZero() bool {
 	return ids.UserID == "" && ids.Email == ""
 }
 
-// Identifiers interface for Entity Identifiers.
-type Identifiers interface {
-	CombinedIdentifiers() *CombinedIdentifiers
-}
-
-// CombinedIdentifiers returns the EntityIdentifiers as a CombinedIdentifiers type.
-func (ids EntityIdentifiers) CombinedIdentifiers() *CombinedIdentifiers {
-	return &CombinedIdentifiers{EntityIdentifiers: []*EntityIdentifiers{&ids}}
-}
-
-// Identifiers returns the actual Identifiers inside the oneof.
-func (ids EntityIdentifiers) Identifiers() Identifiers {
-	switch oneof := ids.Ids.(type) {
-	case *EntityIdentifiers_ApplicationIDs:
-		return oneof.ApplicationIDs
-	case *EntityIdentifiers_ClientIDs:
-		return oneof.ClientIDs
-	case *EntityIdentifiers_DeviceIDs:
-		return oneof.DeviceIDs
-	case *EntityIdentifiers_GatewayIDs:
-		return oneof.GatewayIDs
-	case *EntityIdentifiers_OrganizationIDs:
-		return oneof.OrganizationIDs
-	case *EntityIdentifiers_UserIDs:
-		return oneof.UserIDs
-	default:
-		panic("missed oneof type in EntityIdentifiers.Identifiers()")
-	}
-}
-
-// IDString returns the ID string of the Identifiers inside the oneof.
-func (ids EntityIdentifiers) IDString() string {
-	switch oneof := ids.Ids.(type) {
-	case *EntityIdentifiers_ApplicationIDs:
-		return oneof.ApplicationIDs.GetApplicationID()
-	case *EntityIdentifiers_ClientIDs:
-		return oneof.ClientIDs.GetClientID()
-	case *EntityIdentifiers_DeviceIDs:
-		return fmt.Sprintf("%s.%s", oneof.DeviceIDs.GetApplicationID(), oneof.DeviceIDs.GetDeviceID())
-	case *EntityIdentifiers_GatewayIDs:
-		return oneof.GatewayIDs.GetGatewayID()
-	case *EntityIdentifiers_OrganizationIDs:
-		return oneof.OrganizationIDs.GetOrganizationID()
-	case *EntityIdentifiers_UserIDs:
-		return oneof.UserIDs.GetUserID()
-	default:
-		panic("missed oneof type in EntityIdentifiers.IDString()")
-	}
-}
-
-// EntityType returns the entity type for the Identifiers inside the oneof.
-func (ids EntityIdentifiers) EntityType() string {
-	switch ids.Ids.(type) {
-	case *EntityIdentifiers_ApplicationIDs:
-		return "application"
-	case *EntityIdentifiers_ClientIDs:
-		return "client"
-	case *EntityIdentifiers_DeviceIDs:
-		return "end device"
-	case *EntityIdentifiers_GatewayIDs:
-		return "gateway"
-	case *EntityIdentifiers_OrganizationIDs:
-		return "organization"
-	case *EntityIdentifiers_UserIDs:
-		return "user"
-	default:
-		panic("missed oneof type in EntityIdentifiers.EntityType()")
-	}
-}
-
-// EntityIdentifiers returns the ApplicationIdentifiers as EntityIdentifiers.
-func (ids ApplicationIdentifiers) EntityIdentifiers() *EntityIdentifiers {
-	return &EntityIdentifiers{Ids: &EntityIdentifiers_ApplicationIDs{
-		ApplicationIDs: &ids,
-	}}
-}
-
-// CombinedIdentifiers returns the ApplicationIdentifiers as CombinedIdentifiers.
-func (ids ApplicationIdentifiers) CombinedIdentifiers() *CombinedIdentifiers {
-	return ids.EntityIdentifiers().CombinedIdentifiers()
-}
-
-// EntityIdentifiers returns the ClientIdentifiers as EntityIdentifiers.
-func (ids ClientIdentifiers) EntityIdentifiers() *EntityIdentifiers {
-	return &EntityIdentifiers{Ids: &EntityIdentifiers_ClientIDs{
-		ClientIDs: &ids,
-	}}
-}
-
-// CombinedIdentifiers returns the ClientIdentifiers as CombinedIdentifiers.
-func (ids ClientIdentifiers) CombinedIdentifiers() *CombinedIdentifiers {
-	return ids.EntityIdentifiers().CombinedIdentifiers()
-}
-
-// EntityIdentifiers returns the EndDeviceIdentifiers as EntityIdentifiers.
-func (ids EndDeviceIdentifiers) EntityIdentifiers() *EntityIdentifiers {
-	return &EntityIdentifiers{Ids: &EntityIdentifiers_DeviceIDs{
-		DeviceIDs: &ids,
-	}}
-}
-
-// CombinedIdentifiers returns the EndDeviceIdentifiers as CombinedIdentifiers.
-func (ids EndDeviceIdentifiers) CombinedIdentifiers() *CombinedIdentifiers {
-	return ids.EntityIdentifiers().CombinedIdentifiers()
-}
-
-// EntityIdentifiers returns the GatewayIdentifiers as EntityIdentifiers.
-func (ids GatewayIdentifiers) EntityIdentifiers() *EntityIdentifiers {
-	return &EntityIdentifiers{Ids: &EntityIdentifiers_GatewayIDs{
-		GatewayIDs: &ids,
-	}}
-}
-
-// CombinedIdentifiers returns the GatewayIdentifiers as CombinedIdentifiers.
-func (ids GatewayIdentifiers) CombinedIdentifiers() *CombinedIdentifiers {
-	return ids.EntityIdentifiers().CombinedIdentifiers()
-}
-
-// EntityIdentifiers implements returns theOrganizationIdentifiers as EntityIdentifiers.
-func (ids OrganizationIdentifiers) EntityIdentifiers() *EntityIdentifiers {
-	return &EntityIdentifiers{Ids: &EntityIdentifiers_OrganizationIDs{
-		OrganizationIDs: &ids,
-	}}
-}
-
-// CombinedIdentifiers returns the OrganizationIdentifiers as CombinedIdentifiers.
-func (ids OrganizationIdentifiers) CombinedIdentifiers() *CombinedIdentifiers {
-	return ids.EntityIdentifiers().CombinedIdentifiers()
-}
-
 // OrganizationOrUserIdentifiers returns the OrganizationIdentifiers as *OrganizationOrUserIdentifiers.
 func (ids OrganizationIdentifiers) OrganizationOrUserIdentifiers() *OrganizationOrUserIdentifiers {
 	return &OrganizationOrUserIdentifiers{Ids: &OrganizationOrUserIdentifiers_OrganizationIDs{
 		OrganizationIDs: &ids,
 	}}
-}
-
-// EntityIdentifiers returns the UserIdentifiers as EntityIdentifiers.
-func (ids UserIdentifiers) EntityIdentifiers() *EntityIdentifiers {
-	return &EntityIdentifiers{Ids: &EntityIdentifiers_UserIDs{
-		UserIDs: &ids,
-	}}
-}
-
-// CombinedIdentifiers returns the UserIdentifiers as CombinedIdentifiers.
-func (ids UserIdentifiers) CombinedIdentifiers() *CombinedIdentifiers {
-	return ids.EntityIdentifiers().CombinedIdentifiers()
 }
 
 // OrganizationOrUserIdentifiers returns the UserIdentifiers as *OrganizationOrUserIdentifiers.
@@ -212,40 +69,11 @@ func (ids UserIdentifiers) OrganizationOrUserIdentifiers() *OrganizationOrUserId
 	}}
 }
 
-// Identifiers returns the OrganizationOrUserIdentifiers as Identifiers.
-func (ids OrganizationOrUserIdentifiers) Identifiers() Identifiers {
-	switch oneof := ids.Ids.(type) {
-	case *OrganizationOrUserIdentifiers_OrganizationIDs:
-		return oneof.OrganizationIDs
-	case *OrganizationOrUserIdentifiers_UserIDs:
-		return oneof.UserIDs
-	default:
-		panic("missed oneof type in OrganizationOrUserIdentifiers.Identifiers()")
-	}
-}
-
-// EntityIdentifiers returns the OrganizationOrUserIdentifiers as EntityIdentifiers.
-func (ids OrganizationOrUserIdentifiers) EntityIdentifiers() *EntityIdentifiers {
-	switch oneof := ids.Ids.(type) {
-	case *OrganizationOrUserIdentifiers_OrganizationIDs:
-		return oneof.OrganizationIDs.EntityIdentifiers()
-	case *OrganizationOrUserIdentifiers_UserIDs:
-		return oneof.UserIDs.EntityIdentifiers()
-	default:
-		panic("missed oneof type in OrganizationOrUserIdentifiers.EntityIdentifiers()")
-	}
-}
-
-// CombinedIdentifiers returns the OrganizationOrUserIdentifiers as CombinedIdentifiers.
-func (ids OrganizationOrUserIdentifiers) CombinedIdentifiers() *CombinedIdentifiers {
-	return ids.EntityIdentifiers().CombinedIdentifiers()
-}
-
 // CombineIdentifiers merges the identifiers of the multiple entities.
 func CombineIdentifiers(ids ...Identifiers) *CombinedIdentifiers {
 	combined := &CombinedIdentifiers{}
 	for _, id := range ids {
-		combined.EntityIdentifiers = append(combined.EntityIdentifiers, id.CombinedIdentifiers().GetEntityIdentifiers()...)
+		combined.EntityIdentifiers = append(combined.EntityIdentifiers, id.EntityIdentifiers())
 	}
 	return combined
 }
